#28 example unit test
Removes _GSON_'s reliance on _ApplicationHolder_ and instead allows the converters plugin to autowire the application context. Unfortunately this is only done on the domain class implementation of `asType` so I've had to implement a rather clumsy workaround so that it will work with standard objects. This won't work properly if there are domain instances nested inside a top level non-domain object as it's not using the plugin's `GsonBuilder` instance.